+++ ArtemisKit/Sources/Messages/Services/MessagesService/MessagesServiceImpl+SavedPosts.swift
@@ -12,7 +12,7 @@ extension MessagesServiceImpl {
     struct AddSavedPostRequest: APIRequest {
         typealias Response = RawResponse
 
-        let postId: Int
+        let postId: Int64
         let postType: PostType
 
         var method: HTTPMethod {
@@ -24,7 +24,7 @@ extension MessagesServiceImpl {
         }
     }
 
-    func addSavedPost(with postId: Int, of type: PostType) async -> NetworkResponse {
+    func addSavedPost(with postId: Int64, of type: PostType) async -> NetworkResponse {
         let result = await client.sendRequest(AddSavedPostRequest(postId: postId, postType: type))
 
         switch result {
@@ -64,7 +64,7 @@ extension MessagesServiceImpl {
     struct UpdateSavedPostStatusRequest: APIRequest {
         typealias Response = RawResponse
 
-        let postId: Int
+        let postId: Int64
         let postType: PostType
         let status: SavedPostStatus
 
@@ -77,7 +77,7 @@ extension MessagesServiceImpl {
         }
     }
 
-    func updateSavedPostStatus(for postId: Int, with type: PostType, status: SavedPostStatus) async -> NetworkResponse {
+    func updateSavedPostStatus(for postId: Int64, with type: PostType, status: SavedPostStatus) async -> NetworkResponse {
         let result = await client.sendRequest(UpdateSavedPostStatusRequest(postId: postId, postType: type, status: status))
 
         switch result {
@@ -91,7 +91,7 @@ extension MessagesServiceImpl {
     struct DeleteSavedPostRequest: APIRequest {
         typealias Response = RawResponse
 
-        let postId: Int
+        let postId: Int64
         let postType: PostType
 
         var method: HTTPMethod {
@@ -103,7 +103,7 @@ extension MessagesServiceImpl {
         }
     }
 
-    func deleteSavedPost(with postId: Int, of type: PostType) async -> NetworkResponse {
+    func deleteSavedPost(with postId: Int64, of type: PostType) async -> NetworkResponse {
         let result = await client.sendRequest(DeleteSavedPostRequest(postId: postId, postType: type))
 
         switch result {
